onTaskExecutionCompleted imp

+++ src/main/java/pl/edu/uj/cluster/delegation/FSMBasedDelegationHandler.java
@@ -144,7 +144,7 @@ public class FSMBasedDelegationHandler implements DelegationHandler {
      */
     private boolean delegateTask(Node destinationNode, WorkerPoolTask task) {
         String destinationNodeId = destinationNode.getNodeId();
-        int taskId = task.getTaskId();
+        long taskId = task.getTaskId();
         if (task.isExternal()) {
             ExternalTask externalTask = (ExternalTask) task;
             if (!externalTaskRegistry.remove(externalTask)) {
