+++ src/main/java/pl/edu/uj/cluster/delegation/SimpleDelegationHandler.java
@@ -153,7 +153,7 @@ public class SimpleDelegationHandler implements DelegationHandler {
      */
     private boolean delegateTask(Node destinationNode, WorkerPoolTask task) {
         String destinationNodeId = destinationNode.getNodeId();
-        int taskId = task.getTaskId();
+        long taskId = task.getTaskId();
         if (task.isExternal()) {
             ExternalTask externalTask = (ExternalTask) task;
             if (!externalTaskRegistry.remove(externalTask)) {
