+++ src/main/java/pl/edu/uj/cluster/task/DelegatedTask.java
@@ -34,7 +34,7 @@ public class DelegatedTask implements WorkerPoolTask {
     }
 
     @Override
-    public int getTaskId() {
+    public long getTaskId() {
         return task.getTaskId();
     }
